+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
-## 1.3.0 (November 3, 2021)
+## 2.0.0 (November 12, 2021)
 
 Add support for npm
+
+- Allow npm commands to be run with `run`
+- Alias `add` to `install` when using npm
+- Alias `remove` to `uninstall` when using npm
 
 ## 1.2.0 (October 31, 2021)